% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crossings.R
\name{combinacoes}
\alias{combinacoes}
\title{Combinações formadas nos cruzamentos}
\usage{
combinacoes(base.de.dados, cruzamento)
}
\arguments{
\item{base.de.dados}{Objeto do tipo data.frame contendo a base de dados dos genótipos.}

\item{cruzamento}{Objeto do tipo data.frame especificando os cruzamentos realizados.}
}
\value{
Um objeto do tipo data.frame contendo o número de identificação dos genótipos mãe e pai, o nome do 
cruzamento e o número de sementes.
}
\description{
Identificar as combinações formadas durante os cruzamentos de genótipos.
}
\details{
\subsection{Base de dados}{
A base de dados consiste em uma tabela com três colunas: \code{genotipo_id}, com o número de identificação 
único de cada genótipo, \code{genotipo_nome}, com o nome atribuído à cada genótipo, e 
\code{cruzamento_fonte}, com o cruzamento que deu origem à cada genótipo.
}

\subsection{Cruzamentos}{
Os cruzamentos são organizados em uma tabela com três colunas: \code{mae_id}, com o número de identificação
único dos genótipos usados como mães, \code{pai_id}, com o número de identificação único dos genótipos 
usados como pais, e \code{sementes_numero}, com o número de sementes.
}
}
\note{
Tome cuidado com o sistema de codificação de caracteres usados ao salvar os arquivos da base de dados
e cruzamentos. Dê preferência ao sistema UTF-8. No MS Excel use \emph{Arquivo} > \emph{Salvar como} > 
\emph{CSV UTF-8}.
}
\examples{
\dontrun{
tmp <- combinacoes(
  base.de.dados = read.csv(
    'data/database.csv', sep = ",", header = TRUE, stringsAsFactors = FALSE, na.strings = "", 
    encoding = "UTF-8"),
  cruzamento = read.csv(
    'data/cruzas.csv', sep = ",", header = TRUE, stringsAsFactors = FALSE, encoding = "UTF-8"))
tmp
write.csv(tmp, "data/combinations.csv")
}
}
\author{
Alessandro Samuel-Rosa \email{alessandrosamuelrosa@gmail.com}
}
